feat: Add shortcuts without opening new window
Adding shortcuts was cumbersome before. One had to click the button
for the character they wanted to add a shortcut to, open a new window,
type the shortcut, and save. This would have to be repeated for each
character. Now, all characters are displayed on a new line instead with a
text entry box. The user can enter as many shortcuts as desired and then
apply them. This could still cause annoyance if there are many special
characters to scroll through, but this will do for now.